% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{compute_maximal_partial_clique}
\alias{compute_maximal_partial_clique}
\title{Compute the largest partial clique}
\usage{
compute_maximal_partial_clique(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{The adjacency matrix representing the graph.}

\item{alpha}{The required edge density for the partial clique, must be between 0 and 1.}
}
\value{
A list containing:
\item{node_ids}{A vector containing the indices of the vertices in the computed maximal partial clique.}
\item{edge_density}{The edge density of the computed maximal partial clique.}
}
\description{
This function computes the largest partial clique given an adjacency matrix and a required edge density alpha.
}
\examples{
# Example usage:
# Define an adjacency matrix
adj_mat <- matrix(c(1, 1, 1, 1, 1,
                    1, 1, 1, 1, 1,
                    1, 1, 1, 1, 1,
                    1, 1, 1, 1, 1,
                    1, 1, 1, 1, 1), nrow = 5, byrow = TRUE)

# Set the required edge density
alpha <- 0.7

# Compute the maximal partial clique
compute_maximal_partial_clique(adj_mat, alpha)

}
