% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a random adjacency matrix with a partial clique}
\usage{
generate_partial_clique(n, clique_fraction, clique_edge_density)
}
\arguments{
\item{n}{The total number of nodes in the graph.}

\item{clique_fraction}{The fraction of nodes (of the n nodes) that are part of the partial clique.}

\item{clique_edge_density}{The edge density among the nodes in the clique.}
}
\value{
An adjacency matrix representing the generated graph.
}
\description{
This function generates a random adjacency matrix with a partial clique,
where a specified fraction of nodes forms the partial clique, and the
specified edge density applies within the clique.
}
\examples{
n <- 8
clique_fraction <- 0.5
clique_edge_density <- 0.8
random_adj_matrix <- generate_partial_clique(n, clique_fraction, clique_edge_density)
print(random_adj_matrix)

}
