% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw4_simulation.R
\name{hw4_simulation}
\alias{hw4_simulation}
\title{Simulation for computing maximum partial clique in various implementations}
\usage{
hw4_simulation(
  n_vec,
  clique_fraction_vec,
  clique_edge_density_vec,
  alpha_vec,
  trials,
  imp_numbers
)
}
\arguments{
\item{n_vec}{vector of inputs for number of nodes in adjacency matrix generation}

\item{clique_fraction_vec}{vector of inputs for clique fraction in adjacency matrix generation}

\item{clique_edge_density_vec}{vector of inputs for edge density in adjacency matrix generation}

\item{alpha_vec}{vector of alpha values to use when computing maximum partial clique}

\item{trials}{number of trials per parameterization of adjacency matrix}

\item{imp_numbers}{vector of implementations of compute_maximum_partial_clique}
}
\value{
a list with dataframe with implementation number and number of times each implementation calculated the largest valid partial clique and the number of tests run
}
\description{
This function takes in inputs to generate random adjacency matrices with varying number of nodes, partial clique fractions, and edge densities.
These matrices are used to test which implementations of compute maximum partial clique return the largest valid clique most frequently.
}
\examples{
n_vec <- c(15)
clique_fraction_vec <- c(0.6, 0.9)
clique_edge_density_vec <- c(0.6, 0.9)
alpha_vec <- c(0.75, 0.95)
trials <- 1
imp_numbers <- c(23)

hw4_simulation(n_vec, clique_fraction_vec, clique_edge_density_vec, alpha_vec, trials, imp_numbers)
}
